changed

+++ cart_page/deleteorder.php
@@ -22,7 +22,14 @@ if($_SERVER['REQUEST_METHOD']=="POST"){
     // $sql = "DELETE FROM PRODUCT_ORDER WHERE ORDER_ID = '$order_id'";
     // $arr = oci_parse($conn, $sql);
     // oci_execute($arr);
-    if(isset($order_id[0])){
+    // echo "NEE";
+    $sql = "SELECT COLLECTION_SLOT.SLOT_ID, COLLECTION_SLOT.STATUS FROM COLLECTION_SLOT, PRODUCT_ORDER WHERE COLLECTION_SLOT.SLOT_ID = PRODUCT_ORDER.FK_SLOT_ID AND PRODUCT_ORDER.FK_CART_ID = '$cart_id' AND COLLECTION_SLOT.STATUS = 1";
+    $arr = oci_parse($conn, $sql);
+    oci_execute($arr);
+    $slotidarr = oci_fetch_array($arr);
+    $slotid = $slotidarr[0];
+    $slotstatus = $slotidarr[1];
+    if(isset($order_id[0]) && $slotstatus == 1){
         $order = $order_id[0];
         $sql = "DELETE FROM ORDERED_PRODUCT WHERE ORDER_ID = '$order' AND PRODUCT_ID = '$pid'";
         $arr = oci_parse($conn, $sql);
@@ -34,10 +41,16 @@ if($_SERVER['REQUEST_METHOD']=="POST"){
         while($rows = oci_fetch_array($arr)){
             $count++;
         }
-        if($count == 0){
+        if($count == 0 && $slotstatus == 1){
+           
             $sql = "DELETE FROM PRODUCT_ORDER WHERE ORDER_ID = '$order'";
             $arr = oci_parse($conn, $sql);
             oci_execute($arr);
+            $sql = "DELETE FROM COLLECTION_SLOT WHERE SLOT_ID = '$slotid'";
+            $arr = oci_parse($conn, $sql);
+            oci_execute($arr);
+            // echo "HOGAYA";
+            echo json_encode([true]);
         }
     }
